+++ lib/test_prof/recipes/active_record_shared_connection.rb
@@ -65,6 +65,13 @@ module TestProf
 end
 
 ActiveSupport.on_load(:active_record) do
+  if ::ActiveRecord::Base.connection.pool.respond_to?(:lock_thread=)
+    TestProf.log :warn, "You activated ActiveRecordSharedConnection patch for the Rails version,\n" \
+                        "which has a built-in support for the same functionality.\n" \
+                        "Consider removing it, 'cause this could result in unexpected behaviour.\n\n" \
+                        "Read more in the docs: https://test-prof.evilmartians.io/#/active_record_shared_connection"
+  end
+
   TestProf::ActiveRecordSharedConnection.enable!
   ActiveRecord::Base.singleton_class.prepend TestProf::ActiveRecordSharedConnection::Ext
 end