+++ lib/test_prof/stack_prof/rspec.rb
@@ -1,14 +1,17 @@
 # frozen_string_literal: true
 
-unless "".respond_to?(:parameterize)
-  require "test_prof/ext/string_parameterize"
-  using TestProf::StringParameterize
-end
+require "test_prof/utils/rspec"
 
 module TestProf
   module StackProf
     # Reporter for RSpec to profile specific examples with StackProf
     class Listener # :nodoc:
+      class << self
+        attr_accessor :report_name_generator
+      end
+
+      self.report_name_generator = Utils::RSpec.method(:example_to_filename)
+
       NOTIFICATIONS = %i[
         example_started
         example_finished
@@ -24,7 +27,7 @@ module TestProf
         return unless notification.example.metadata[:sprof_report] == false
 
         TestProf::StackProf.dump(
-          ::RSpec::Core::Metadata.id_from(notification.example.metadata).parameterize
+          self.class.report_name_generator.call(notification.example)
         )
       end